refactor: reorganize link checker module and update imports
- Move link_checker_example.py to digital_assistant_first/utils/
- Create a new link_checker.py with simplified implementation
- Update interface.py import to use new link_checker module
- Remove get_invalid_links tool from corrector agent
- Adjust model to gpt-4o-mini for consistency
