+++ digital_assistant_first/utils/link_checker.py
@@ -1,23 +1,22 @@
-from pydantic_ai import Agent, RunContext
+from pydantic_ai import Agent
 from pydantic import BaseModel, Field
-from typing import List, Dict
+from typing import List
 import httpx
 from dotenv import load_dotenv
-import asyncio
 import logfire
 
-logfire.configure()
+# logfire.configure()
 
 load_dotenv()
 
 
 class LinkStatus(BaseModel):
-    link: str
-    status: bool
+    link: str = Field(description="The link to check")
+    status: bool = Field(description="The status of the link")
 
 
 class LinkStatusList(BaseModel):
-    links: List[LinkStatus]
+    links: List[LinkStatus] = Field(description="The list of links to check")
 
 
 link_checker = Agent(
@@ -43,8 +42,6 @@ corrector = Agent(
         "3. Make the minimal changes necessary to maintain text coherence after removing invalid links. "
         "4. Do not modify, improve, or change any other parts of the text. "
         "5. Keep all valid links exactly as they appear in the original text."
-        "Use the get_invalid_links tool to get the invalid links."
-        "Respond only with the corrected text, nothing else."
     ),
     deps_type=LinkStatusList,
     result_type=str,
@@ -66,20 +63,3 @@ async def check_link_list(list_of_links: List[str]) -> LinkStatusList:
             except httpx.RequestError:
                 status.append(LinkStatus(link=link, status=False))
     return LinkStatusList(links=status)
-
-
-@corrector.tool
-async def get_invalid_links(ctx: RunContext[LinkStatusList]) -> List[str]:
-    return [link.link for link in ctx.deps.links if not link.status]
-
-
-async def main():
-    text = "Here are some useful resourses: https://www.google.com, https://www.youtube.com, https://yandex.ru/maps/org/yozh_ustritsa/52393193425/"
-    result = await link_checker.run(text)
-    print(result.data)
-    corrected_text = await corrector.run(text, deps=result.data)
-    print(corrected_text.data)
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
